dns: change behavior for rc-manager=symlink to preserve "/etc/resolv.conf" as file
The purpose of "rc-manager=symlink" is so that the administrator can point
the "/etc/resolv.conf" as a symlink to a certain file, and thus indicating
that a certain party is responsible to manage resolv.conf, while others
should stay away from it.
For example, systemd-resolved never touches "/etc/resolv.conf", but
expects the admin to setup the symlink appropriately. It also recognizes
whether the symlink points to it's own resolv.conf copy or to another
component.

Previously, rc-manager=symlink would always replace a regular file
with a symlink to "/var/run/NetworkManager/resolv.conf". If
"/etc/resolv.conf" is already a symlink somewhere else, it would not
touch it, with the exception that if "/etc/resolv.conf" points to
"/var/run/NetworkManager/resolv.conf", it would replace the symlink,
merely to raise inotify events.

Change behavior so now if "/etc/resolv.conf" is already a file, keep
it as file.
This means, if you have multiple components that don't care, everybody
can write the "/etc/resolv.conf" (as file) and there is no clear
expressed responsibility.
It was wrong that NetworkManager tries to convert a file to a symlink,
this should be reserved to the admin. Instead, NetworkManager should
accept that the intent is unclear and preserve it as file. It's up to the
admin to replace the symlink to somewhere else (to keep NM off), or to
point it to "/var/run/NetworkManager/resolv.conf", to show the explicit
intent.

The wrong behavior causes dangling symlinks when somebody disables
NetworkManager.

https://bugzilla.redhat.com/show_bug.cgi?id=1367551